gun/rifle: merge rifle control in TRX
