+++ src/tr1/meson.build
@@ -99,7 +99,6 @@ sources = [
   init,
   'game/creature.c',
   'game/game_string.c',
-  'game/gun/gun_rifle.c',
   'game/objects/common.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
