--- src/tr2/game/gun/gun_rifle.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <libtrx/game/gun/rifle.h>
-
-void Gun_Rifle_Animate(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_Ready(LARA_GUN_TYPE weapon_type);
